added Multi Copy Entry
removed '.tmp' file from git
