+++ test/tests/test-book-store.cc
@@ -1,43 +1,48 @@
-// -- test include
+
 #include <UnitTesting.h>
-/*
-// -- marlin include
+#include <typeinfo>
+
+#include "marlin/book/Hist.h"
+#include "marlin/book/Handle.h"
 #include "marlin/book/BookStore.h"
-#include "marlin/book/Flags.h"
-#include "marlin/book/hist.h"
+#include "ROOT/RHistData.hxx" 
+#include "ROOT/RHist.hxx"
 
-using namespace marlin::book::Types;
+#include <memory>
 
-int main(int, char**) {
-	marlin::test::UnitTest test (" BookStore ");
-
-	marlin::book::BookStore store{};
-	auto hnd1 = store.book<RH1D, RAxisConfig>("name", "path", 0, 2, marlin::book::Flag_t{}, {"x", 2, 1.0, 0.0});
-	auto hnd2 = store.book<RH1D, RAxisConfig>("name", "path", 1, 2, marlin::book::Flag_t{}, {"x", 2, 1.0, 0.0});
-	hnd1->fill({1}, 1);
-	hnd2->fill({1}, 1);
-	hnd1->fill({0}, 1);
-	std::cout << hnd1.get().GetBinContent({1}) << " " << hnd2.get().GetBinContent({0}) << '\n';
-	test.test(
-		"Concurrent Test",
-		hnd1.get().GetBinContent({1}) == 2
-		&& hnd2.get().GetBinContent({0}) == 1);
-	
-	auto hnd3 = store.book<RH1D, RAxisConfig>("name2", "path", 0, 2, marlin::book::Flags::Book::MultiInstance, {"x", 2, 1.0, 0.0});
-	auto hnd4 = store.book<RH1D, RAxisConfig>("name2", "path", 1, 2, marlin::book::Flags::Book::MultiInstance, {"x", 2, 1.0, 0.0});
-	hnd3->fill({1}, 1);
-	hnd3->fill({1}, 1);
-	hnd4->fill({0}, 1);
-	test.test(
-		"Parallel Test", 
-		hnd3.get().GetBinContent({0}) == 1
-		&& hnd4.get().GetBinContent({1}) == 2); 
-	return 0;
-}*/
-
-int main(int, char**) {
-	marlin::test::UnitTest test( " EmptyTest " );	
-	return 0;
+template<typename T>
+void MergeaHist(const std::shared_ptr<T>& dst, const std::shared_ptr<T>& src) {
+	ROOT::Experimental::Add(*dst, *src);
 }
 
+using namespace marlin::book;
+using namespace ROOT::Experimental;
 
+int main(int, char**) {
+	marlin::test::UnitTest test(" MemFillerTest ");
+	BookStore store{};
+
+	{
+
+		EntrySingle entry = store.book<RH<float, 1>, RAxisConfig>("test", "path", {"a", 3, 1.0, 2.0}) ;	
+		auto hnd = entry.handle();
+		hnd.fill({0}, 1);
+		auto hist = hnd.get();
+		test.test("Single Hist Filling", hist.GetEntries() == 1);
+
+	}{
+
+		EntryMultiCopy entry = store.bookMultiCopy<RH<int, 1>, RAxisConfig>(2, "test", "path", {"a", 3, 1.0, 2.0});
+		auto hnd = entry.handle(0);
+		hnd.fill({0}, 1);
+
+		auto hnd2 = entry.handle(1);
+		hnd2.fill({0}, 1);
+
+		auto hist = hnd.get();
+		test.test("MultiCopd Hist Filling", hist.GetBinContent({0})== 2);
+
+	}
+
+	return 0;
+}
